requirements/test.txt
coverage>=5.0.4
coveralls>=2.0.0
parameterized>=0.7.4
python-dotenv>=0.14.0